Split controllers into two separate packages.
