+++ src/main/java/org/example/springboot_labb2/controller/rest/UserController.java
@@ -1,4 +1,4 @@
-package org.example.springboot_labb2.controller;
+package org.example.springboot_labb2.controller.rest;
 
 import org.example.springboot_labb2.exception.ResourceNotFoundException;
 import org.example.springboot_labb2.repository.UserRepository;
